==> jupyter_notebooks/01_Data_Collection.ipynb <==
# **Notebook 1: Data Collection**

 
 
## Objectives

* Fetch data from Kaggle and save as raw data.

* Load and inspect the data and save under outputs/datasets/collection

* Push files to github repository.

## Inputs

* Kaggle JSON - authentication token.

* Downloaded the recommended house price dataset from [Kaggle](https://www.kaggle.com/datasets/codeinstitute/housing-prices-data)

## Outputs

* The Kaggle files were unzipped to reveal the following files for input:
* inputs/datasets/raw/house-price-20211124T154130Z-001/house-price/house_prices_records.csv
* inputs/datasets/raw/house-price-20211124T154130Z-001/house-price/inherited_houses.csv
* Generate Dataset: input/datasets/house-price/house_prices_records and input/datasets/house-price/inherited_houses

## Additional Comments

* The Data understanding within Crisp-DM methodology is shown in this notebook.

* Walkthrough Project 2 has formed the guidelines for this notebook in terms of undertaking data collection.

---

# Change working directory

Here we will change the working directory from its current folder to its parent folder.
* We access the current directory with the command os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory

* os.path.dirname() gets the parent directory

* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Here we will confirm the new current directory.

current_dir = os.getcwd()
current_dir

---

# Fetch data from Kaggle

* Required to install the Kaggle library hereafter.

Section 1 content

 ! pip3 install kaggle==1.5.12

* Your are required to create a Kaggle account and a kaggle.json file (to be downloaded) in order to allow for the data download to work.

* The Kaggle.json file hosts authentication token, which is required to authenticate a data download from Kaggle itself, one can be requested from your Kaggle account settings under API's (create new token).

* We can then copy the downloaded kaggle.json file into the root directory of the project repository.

* Next, we set the Kaggle environment variable and permission to the token.

---

# Using your key

* Once you have downloaded your API key from the kaggle website, please place the file in the root directory for your project. You would just need to drag and drop.

* Thereafter, run the below cell to ensure the correct permissions are assigned for handling the file.

import os
os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

---

# Retrieving the dataset path from the Kaggle URL

* We are using the Kaggle dataset called House Prices.

* Remember to copy the section of the URL on Kaggle that comes after "https://www.kaggle.com/"

* We then define the dataset and it's destination folder after download.

KaggleDatasetPath = "codeinstitute/housing-prices-data"
DestinationFolder = "inputs/datasets/raw"   
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

* We now have to unzip the downloaded file, delete the zip file and delete the kaggle.json file.

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json 

! pip3 uninstall -y kaggle==1.5.12

---

# Import packages & set environment variables

import pandas as pd
pd.options.display.max_columns = None
pd.options.display.max_rows = None

---

# Load and Inspect the House Price Records

* We can read the house_prices_records dataset csv file into a Pandas dataframe.

df = pd.read_csv(f"inputs/datasets/raw/house-price-20211124T154130Z-001/house-price/house_prices_records.csv")
print(df.shape)
df.head()

---

# Dataframe Summary
* We are also able to read the dataframe summary by calling the method .info() on the dataframe object, but the snippet below reads this output into it's own dataframe for readability purposes.

df.info()

* From above there is no id field to show data uniqueness and thus there is no need to check for duplicate data.

---

# Load and Inspect Inherited House Records

* Read the inherited_houses dataset csv file into a Pandas dataframe.

df_inherited = pd.read_csv(f"inputs/datasets/raw/house-price-20211124T154130Z-001/house-price/inherited_houses.csv")
print(df_inherited.shape)
df_inherited

---

# Conclusions and Steps to Follow

* There is a difference between the house price and inherited house datasets.

* Some features in the house price dataset are in int while others are in float type.

* However, the above should not affect the analysis of the data.

* Salesprice is only found in the house price dataset in integer form.

* We can now proceed towards cleaning our data.
